+++ src/builtins/fish_key_reader.rs
@@ -22,7 +22,7 @@ use crate::{
         terminal_protocols_enable_ifn, Capability, CharEvent, ImplicitEvent, InputEventQueue,
         InputEventQueuer, KeyEvent, KITTY_KEYBOARD_SUPPORTED,
     },
-    key::{char_to_symbol, Key},
+    key::{char_to_symbol, Key, Modifiers},
     nix::isatty,
     panic::panic_handler,
     print_help::print_help,
@@ -102,9 +102,33 @@ fn process_input(streams: &mut IoStreams, continuous_mode: bool, verbose: bool) 
             }
             streams.out.append(L!("\n"));
         }
-        streams
-            .out
-            .append(sprintf!("bind %s 'do something'\n", kevt.key));
+        let mut print_bind_example = |key: &Key, recommended: bool| {
+            streams.out.append(sprintf!(
+                "bind %s 'do something'%s\n",
+                key,
+                if recommended {
+                    " # recommended notation"
+                } else {
+                    ""
+                }
+            ));
+        };
+        let have_shifted_key = kevt.key.shifted_codepoint != '\0';
+        // If we have shift + some other modifier, the lowercase version is the canonical one.
+        let prefer_explicit_shift = kevt.key.modifiers.shift
+            && kevt.key.modifiers != Modifiers::SHIFT
+            && kevt
+                .key
+                .shifted_codepoint
+                .to_lowercase()
+                .eq(Some(kevt.key.codepoint).into_iter());
+        if have_shifted_key {
+            let mut shifted_key = kevt.key.key;
+            shifted_key.modifiers.shift = false;
+            shifted_key.codepoint = kevt.key.shifted_codepoint;
+            print_bind_example(&shifted_key, !prefer_explicit_shift);
+        }
+        print_bind_example(&kevt.key, have_shifted_key && prefer_explicit_shift);
 
         if continuous_mode && should_exit(streams, &mut recent_chars, kevt.key) {
             streams.err.appendln("\nExiting at your request.");